+++ agen5/expPrint.c
@@ -1,9 +1,9 @@
 
 /*
- *  $Id: expPrint.c,v 4.14 2006/09/24 02:57:01 bkorb Exp $
+ *  $Id: expPrint.c,v 4.15 2006/11/27 01:55:17 bkorb Exp $
  *
- *  Time-stamp:        "2006-09-23 19:52:48 bkorb"
- *  Last Committed:    $Date: 2006/09/24 02:57:01 $
+ *  Time-stamp:        "2006-11-26 15:38:58 bkorb"
+ *  Last Committed:    $Date: 2006/11/27 01:55:17 $
  *
  *  The following code is necessary because the user can give us
  *  a printf format requiring a string pointer yet fail to provide
@@ -308,7 +308,7 @@ ag_scm_hide_email( SCM display, SCM eaddr )
 
     pzScan += sprintf( pzScan, zEnd, pzDisp );
 
-    return AG_SCM_STR2SCM( pzRes, (unsigned)(pzScan - pzRes));
+    return AG_SCM_STR2SCM( pzRes, (size_t)(pzScan - pzRes));
 }
